// src/types/ProductType.ts
export interface Category {
  id: number;
  name: string;
  slug: string;
  image: string;
  creationAt: string;
  updatedAt: string;
}

export interface Product {
  id: number;
  title: string;
  slug: string;
  price: number;
  description: string;
  category: Category;
  images: string[];
  creationAt: string;
  updatedAt: string;
}

export interface ProductCardProps {
  id: string;
  badge?: string;
  imageSrc: string;
  imageAlt?: string;
  title: string;
  price: number;
}